portation of product_brand to 8.0
