+++ product_brand/product_brand_view.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  product_brand for OpenERP
-  Copyright (C) 2009 NetAndCo (<http://www.netandco.net>).
-    Authors, Mathieu Lemercier, mathieu@netandco.net, Franck Bret, franck@netandco.net
+  product_brand for Odoo
+  Copyright (C) 2009 NetAndCo <http://www.netandco.net>.
   Copyright (C) 2011 Akretion Benoît Guillot <benoit.guillot@akretion.com>
-  The licence is in the file __openerp__.py
+  Copyright (C) 2014 prisnet.ch Seraphine Lantible <s.lantible@gmail.com> 
+    Contributors, Mathieu Lemercier <mathieu@netandco.net>, Franck Bret <franck@netandco.net>, Seraphine Lantible <s.lantible@gmail.com>
 -->
 <openerp>
     <data>
@@ -71,25 +71,13 @@
             </field>
         </record>
 
-        <record model="ir.ui.view" id="normal_form_product_brand">
-           <field name="name">product.normal.form</field>
-           <field name="model">product.product</field>
-           <field name="inherit_id" ref="product.product_normal_form_view" />
-           <field name="arch" type="xml">
-                <div name="options" position="before">
-                    <label for="product_brand_id" class="oe_edit_only"/>
-                    <h2><field name="product_brand_id"/></h2>
-               </div>
-           </field>
-        </record>
-
-        <record model="ir.ui.view" id="product_template_form_brand_add">
+         <record model="ir.ui.view" id="product_template_form_brand_add">
             <field name="name">product.template.product.form</field>
             <field name="model">product.template</field>
             <field name="inherit_id" ref="product.product_template_form_view" />
             <field name="arch" type="xml">
-                <field name="categ_id" position="after"  >
-                  <field name="product_brand_id" />
+                <field name="name" position="after"  >
+                  <field name="product_brand_id" placeholder="Marque"/>
                 </field>
               </field>
         </record>
